// frontend/src/app/shared/models/creatures/characters/character-page-type.enum.ts
export enum CharacterPageType {
  BASIC = 'BASIC',
  ABILITIES = 'ABILITIES',
  QUICK_ACTIONS = 'QUICK_ACTIONS',
  SPELLS = 'SPELLS',
  SKILLS = 'SKILLS',
  FEATURES = 'FEATURES',
  CONDITIONS = 'CONDITIONS',
  PROFICIENCIES = 'PROFICIENCIES',
  DAMAGE_MODIFIERS = 'DAMAGE_MODIFIERS',
  COMPANIONS = 'COMPANIONS',
  EQUIPMENT = 'EQUIPMENT',
  CHARACTERISTICS = 'CHARACTERISTICS',
  NOTES = 'NOTES'
}

export enum EncounterPageType {
  BASIC = 'BASIC',
  ABILITIES = 'ABILITIES',
  QUICK_ACTIONS = 'QUICK_ACTIONS',
  SPELLS = 'SPELLS',
  SKILLS = 'SKILLS',
  FEATURES = 'FEATURES',
  CONDITIONS = 'CONDITIONS',
  PROFICIENCIES = 'PROFICIENCIES',
  DAMAGE_MODIFIERS = 'DAMAGE_MODIFIERS',
  COMPANIONS = 'COMPANIONS',
  EQUIPMENT = 'EQUIPMENT',
  CHARACTERISTICS = 'CHARACTERISTICS',
  NOTES = 'NOTES'
}